azure-vm/api/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
Pillow>=10.0.0
azure-storage-blob>=12.0.0
azure-identity>=1.15.0
aiofiles>=23.2.0
python-dotenv>=1.0.0

# Azure Functions (for function app deployment only)
azure-functions

# Development/Testing
pytest>=8.0.0
